Introduce rspec-puppet and compile test for each class
This PR adds the setup for rspec-puppet unit tests to the module, and
adds a very simple "does the manifest compile" test for each class. I
wanted to add all the setup in a separate PR so that it didn't get too
bloated.

Puppet/rspec-puppet seem to have a fun bug revolving around Windows file
paths in which Puppet won't correctly detect what kernel you're
intending, and will complain that the Windows file path (which doesn't
start with a `/`) is unqualified. I included a workaround in
spec_helper.rb from a rspec-puppet thread about the issue.

https://github.com/rodjek/rspec-puppet/issues/192

Furthermore, a couple params weren't being passed properly to the
win_config.pp class, so I fixed those as well so the module can compile.